Refactor, extract contents and drop CaseAPIClasses and CaseList
The file CaseAPIClasses contained case classes that hold formats for query results and also for history responses. This is separated into separate classes.
The ones for the query results have moved into the querydb package.
