+++ src/main/scala/org/cafienne/service/akkahttp/cases/history/PlanItemHistory.scala
@@ -1,0 +1,25 @@
+/*
+ * Copyright (C) 2014  Batav B.V.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.cafienne.service.akkahttp.cases.history
+
+import org.cafienne.json.{CafienneJson, Value}
+import org.cafienne.querydb.record.PlanItemHistoryRecord
+
+final case class PlanItemHistory(records: Seq[PlanItemHistoryRecord]) extends CafienneJson {
+  override def toValue: Value[_] = Value.convert(records.map(item => item.toValue))
+}
